requirements.txt
pandas==2.0.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
celery==5.3.6
redis==5.0.1
flower==2.0.1